ses: refactoring
Move ses functions from scsi.c to ses.c and enclosure functions to
enclosure.c.
